# openapi.yaml
openapi: 3.0.3
info:
  title: Mock Order API
  description: |
    Simple mock Order Management API for front-end testing.

    **API Versioning Strategy**:
    - This API uses URL-based versioning
    - Current version: v1 (available at `/api/v1`)
    - Version is embedded in the URL path: `/api/{version}/resource`
    - When breaking changes are introduced, a new version will be released (e.g., `/api/v2`)
    - Clients must specify the version in the URL path when making requests
    - Example: `/api/v1/orders` (v1) vs `/api/v2/orders`

    **Authentication**: All endpoints require Bearer token authentication.
    Include the token in the Authorization header: `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: /api/v1
    description: Version 1 of the API
security:
  - bearerAuth: []

paths:
  /orders:
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOrder"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Search orders by customer and date range
      parameters:
        - in: query
          name: customerId
          required: true
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /orders/{orderId}/status:
    patch:
      summary: Update order status
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [PLACED, SHIPPED, CANCELLED]
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NewOrder:
      type: object
      required: [customerId]
      properties:
        customerId: { type: string }
        placementDate: { type: string, format: date-time }
    Order:
      type: object
      required: [customerId, orderId, status]
      properties:
        orderId: { type: string }
        customerId: { type: string }
        placementDate: { type: string, format: date-time }
        status:
          type: string
          enum: [PLACED, SHIPPED, CANCELLED]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional error details (optional)